+++ main.go
@@ -6,7 +6,7 @@ import (
 )
 
 //	@title			Device Management Service
-//	@version		0.4.125
+//	@version		0.4.126
 //	@description	A dashboard application for computing providers.
 //	@termsOfService	https://nunet.io/tos
 
@@ -17,8 +17,8 @@ import (
 //	@license.name	Apache 2.0
 //	@license.url	http://www.apache.org/licenses/LICENSE-2.0.html
 
-//	@host		localhost:9999
-//	@BasePath	/api/v1
+// @host		localhost:9999
+// @BasePath	/api/v1
 func main() {
 	db.ConnectDatabase()
 	cmd.Execute()